Split Item module from RuneLite module
