+++ src/modules/index.js
@@ -1,5 +1,6 @@
 import appReducer from './app'
 import gitReducer from './git'
+import itemReducer from './item'
 import runeliteReducer from './runelite'
 import tagReducer from './tag'
 import xpReducer from './xp'
@@ -8,6 +9,7 @@ import xpReducer from './xp'
 export default {
   app: appReducer,
   git: gitReducer,
+  item: itemReducer,
   runelite: runeliteReducer,
   tag: tagReducer,
   xp: xpReducer
